cleanup try except for commits

+++ app/uploads.py
@@ -25,7 +25,7 @@ from .models import (
     context_str_to_enum,
     status_str_to_enum,
     reset_all_gqs,
-    ensure_admin,
+    try_sql_commit,
     ensure_all_gqs,
     drop_and_rebuild_tables,
 )
@@ -68,10 +68,7 @@ def delete_all_clusters():
             db.session.add(paper)
     num_deleted = Label.query.filter(Label.is_cluster).delete()
     print(f"delete {num_deleted} cluster labels.")
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
+    if not try_sql_commit():
         msg = "failed in deleting clusters"
         print(msg)
         flash(msg)
@@ -91,10 +88,7 @@ def delete_all_paper_rooms():
             db.session.add(paper)
     num_deleted = Label.query.filter(Label.is_room).delete()
     print(f"delete {num_deleted} cluster labels.")
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
+    if not try_sql_commit():
         msg = "failed in deleting paper rooms"
         print(msg)
         flash(msg)
@@ -108,14 +102,13 @@ def delete_all_labels():
     for label in labels:
         label.tag_papers = []  # empty list
         db.session.add(label)
-    try:
-        db.session.commit()
+    ok = try_sql_commit()
+    if ok:
         num_deleted = Label.query.delete()
-        db.session.commit()
         print(f"deleted {num_deleted} labels")
-    except:
-        db.session.rollback()
-        msg = "failed in delete_all_labels"
+        ok = try_sql_commit()
+    if not ok:
+        msg = "failed commit in delete_all_labels"
         print(msg)
         flash(msg)
     dump_users_papers_and_conflicts("After label deletion")
@@ -125,12 +118,9 @@ def delete_all_labels():
 def delete_all_history():
     dump_users_papers_and_conflicts("Before History deletion")
     num_deleted = History.query.delete()
-    try:
-        db.session.commit()
-        print(f"Deleted {num_deleted} history entries.")
-    except:
-        db.session.rollback()
-        msg = "failed in delete_all_history"
+    print(f"Deleted {num_deleted} history entries.")
+    if not try_sql_commit():
+        msg = "failed commit in delete_all_history"
         print(msg)
         flash(msg)
     dump_users_papers_and_conflicts("After History deletion")
@@ -143,12 +133,9 @@ def delete_non_bbs_history():
     context_bbs = int(HistoryContext.BBS)
     # Note that filter() allows for != (but filter_by does not allow it)
     num_deleted = History.query.filter(History.context_enum != context_bbs).delete()
-    try:
-        db.session.commit()
-        print(f"Deleted {num_deleted} history entries.")
-    except:
-        db.session.rollback()
-        msg = "failed in delete_non_bbs_history"
+    print(f"Deleted {num_deleted} history entries.")
+    if not try_sql_commit():
+        msg = "failed commit in delete_non_bbs_history"
         print(msg)
         flash(msg)
     dump_users_papers_and_conflicts("After non-BBS History deletion")
@@ -160,11 +147,8 @@ def delete_all_summaries():
     for paper in papers:
         paper.summary = ""
         db.session.add(paper)
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed in delete_all_summaries"
+    if not try_sql_commit():
+        msg = "failed commit in delete_all_summaries"
         print(msg)
         flash(msg)
     print(f"Deleted {count} summaries.")
@@ -173,11 +157,8 @@ def delete_all_summaries():
 # 2023?
 def delete_all_uploads():
     num_deleted = FileUpload.query.delete()
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed in delete_all_uploads"
+    if not try_sql_commit():
+        msg = "failed commit in delete_all_uploads"
         print(msg)
         flash(msg)
     print(f"Deleted {num_deleted} file upload entries.")
@@ -185,11 +166,8 @@ def delete_all_uploads():
 
 def delete_all_queries():
     num_deleted = Query.query.delete()
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed in delete_all_queries"
+    if not try_sql_commit():
+        msg = "failed commit in delete_all_queries"
         print(msg)
         flash(msg)
     print(f"Deleted {num_deleted} queries.")
@@ -207,11 +185,8 @@ def insert_query_rows(rows):
         query = Query(name=name, json=json_quote)
         db.session.add(query)
         count += 1
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed to insert query rows"
+    if not try_sql_commit():
+        msg = "failed commit in insert query rows"
         print(msg)
         flash(msg)
         return 0
@@ -239,11 +214,8 @@ def insert_user_rows(rows):
             user.role = roleObj
         db.session.add(user)
         count += 1
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed to insert user rows (possible duplicate email?)"
+    if not try_sql_commit():
+        msg = "failed commit when insert user rows (possible duplicate email?)"
         print(msg)
         flash(msg)
         return 0
@@ -308,11 +280,8 @@ def insert_paper_rows(rows):
             if label and paper:
                 paper.tag_labels.append(label)
                 db.session.add(paper)
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed to insert paper rows (possible duplicate paper id?)"
+    if not try_sql_commit():
+        msg = "failed commit when insert paper rows (possible duplicate paper id?)"
         print(msg)
         flash(msg)
         return 0
@@ -333,11 +302,8 @@ def insert_conflict_rows(rows):
             user.conf_papers.append(paper)
             db.session.add(user)
             count += 1
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed to insert conflict rows"
+    if not try_sql_commit():
+        msg = "failed commit in insert conflict rows"
         print(msg)
         flash(msg)
         return 0
@@ -357,11 +323,8 @@ def insert_summary_rows(rows):
             paper.summary = summary
             db.session.add(paper)
             count += 1
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed to insert summaries"
+    if not try_sql_commit():
+        msg = "failed commit when insert summaries"
         print(msg)
         flash(msg)
         return 0
@@ -388,11 +351,8 @@ def insert_label_rows(rows, label_type):
             paper.tag_labels.append(label)
             db.session.add(paper)
             count += 1
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = f"failed to insert labels of type {label_type}"
+    if not try_sql_commit():
+        msg = f"failed commit when insert labels of type {label_type}"
         print(msg)
         flash(msg)
         return 0
@@ -434,11 +394,8 @@ def insert_people_room_rows(rows):
             person.rooms = rooms
             db.session.add(person)
             count += 1
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = f"failed to insert people rooms"
+    if not try_sql_commit():
+        msg = f"failed commit when insert people rooms"
         print(msg)
         flash(msg)
         return 0
@@ -506,11 +463,8 @@ def insert_chair_score_rows(rows):
         )
         db.session.add(history)
         count += 1
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed to insert chair scores"
+    if not try_sql_commit():
+        msg = "failed commit in insert chair scores"
         print(msg)
         flash(msg)
         return 0
@@ -545,11 +499,8 @@ def insert_history_rows(rows):
         )
         db.session.add(history)
         count += 1
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed to insert history"
+    if not try_sql_commit():
+        msg = "failed commit in insert history"
         print(msg)
         flash(msg)
         return 0
@@ -686,11 +637,8 @@ def read_csv(filename):
     delete_prev_file_uploads(header_type)
     upload = FileUpload(file=header_type, count=count)
     db.session.add(upload)
-    try:
-        db.session.commit()
-    except:
-        db.session.rollback()
-        msg = "failed to add file upload record"
+    if not try_sql_commit():
+        msg = "failed commit in add file upload record"
         print(msg)
     msg = dump_users_papers_and_conflicts("After Upload")
     if header_type == "users":
